+++ tree/node.h
@@ -23,7 +23,7 @@ const char MAXHIGHKEY[] = "infinity";
 
 struct Stdhead {
     uint16_t key_offset;
-    uint8_t key_len;
+    uint16_t key_len;
 } __attribute__((packed));
 
 class Node {
@@ -33,7 +33,7 @@ public:
     char *base;
     uint16_t space_top;
     vector<Node *> ptrs;
-    uint8_t ptr_cnt;
+    uint16_t ptr_cnt;
 
     Item *lowkey;
     Item *highkey;
@@ -46,14 +46,14 @@ public:
 
 struct DB2head {
     uint16_t key_offset;
-    uint8_t key_len;
+    uint16_t key_len;
 } __attribute__((packed));
 
 struct DB2pfxhead {
     uint16_t pfx_offset;
-    uint8_t pfx_len;
-    uint8_t low;
-    uint8_t high;
+    uint16_t pfx_len;
+    uint16_t low;
+    uint16_t high;
 } __attribute__((packed));
 
 class NodeDB2 {
@@ -66,7 +66,7 @@ public:
     uint16_t space_top;
     uint16_t pfx_top;
     vector<NodeDB2 *> ptrs;
-    uint8_t ptr_cnt;
+    uint16_t ptr_cnt;
     NodeDB2 *prev; // Prev node pointer
     NodeDB2 *next; // Next node pointer
     NodeDB2();
@@ -78,8 +78,8 @@ public:
 
 struct MyISAMhead {
     uint16_t key_offset;
-    uint8_t key_len;
-    uint8_t pfx_len;
+    uint16_t key_len;
+    uint16_t pfx_len;
     /* We assume all the prefix are less than 128 Byte, so remove is1Byte in MyISAM
      *  May need to consider the case key/pfx_len in [128, 256)
      */
@@ -114,8 +114,8 @@ struct WThead {
 #else
 struct WThead {
     uint16_t key_offset;
-    uint8_t key_len;
-    uint8_t pfx_len;
+    uint16_t key_len;
+    uint16_t pfx_len;
 } __attribute__((packed));
 #endif
 
@@ -183,8 +183,8 @@ public:
 
 #else
 struct PkBhead {
-    uint8_t pfx_len;
-    uint8_t key_len;
+    uint16_t pfx_len;
+    uint16_t key_len;
     uint8_t pk_len;
     char pk[PKB_LEN + 1];
     uint16_t key_offset;
