+++ tree_disk/node_disk_inline.h
@@ -53,7 +53,7 @@
 
 #define GetHeaderStd(nptr, i) (Stdhead *)(nptr->base + MAX_SIZE_IN_BYTES - (i + 1) * sizeof(Stdhead))
 
-inline void InsertKeyStd(Node *nptr, int pos, const char *k, uint8_t klen) {
+inline void InsertKeyStd(Node *nptr, int pos, const char *k, uint16_t klen) {
     strcpy(BufTop(nptr), k);
     // shift the headers
     for (int i = nptr->size; i > pos; i--) {
@@ -68,7 +68,7 @@ inline void InsertKeyStd(Node *nptr, int pos, const char *k, uint8_t klen) {
 }
 
 // with cutoff
-inline void CopyToNewPageStd(Node *nptr, int low, int high, char *newbase, uint8_t cutoff, uint16_t &top) {
+inline void CopyToNewPageStd(Node *nptr, int low, int high, char *newbase, uint16_t cutoff, uint16_t &top) {
     for (int i = low; i < high; i++) {
         int newidx = i - low;
         Stdhead *oldhead = GetHeaderStd(nptr, i);
@@ -102,7 +102,7 @@ inline void CopyToNewPageStd(Node *nptr, int low, int high, char *newbase, uint8
 
 #define PfxOffset(node, off) (char *)(node->base + MAX_SIZE_IN_BYTES + off)
 
-inline void InsertPfxDB2(NodeDB2 *nptr, int pos, const char *p, uint8_t plen, uint8_t low, uint8_t high) {
+inline void InsertPfxDB2(NodeDB2 *nptr, int pos, const char *p, uint16_t plen, uint16_t low, uint16_t high) {
     char *temp = PfxTop(nptr);
     memcpy(temp, p, sizeof(char) * plen);
     temp[plen] = '\0';
@@ -120,7 +120,7 @@ inline void InsertPfxDB2(NodeDB2 *nptr, int pos, const char *p, uint8_t plen, ui
     nptr->pfx_size += 1;
 }
 
-inline void InsertKeyDB2(NodeDB2 *nptr, int pos, const char *k, uint8_t klen) {
+inline void InsertKeyDB2(NodeDB2 *nptr, int pos, const char *k, uint16_t klen) {
     strcpy(BufTop(nptr), k);
     // shift the headers
     for (int i = nptr->size; i > pos; i--) {
@@ -187,8 +187,8 @@ inline void InsertKeyWT(NodeWT *nptr, int pos, const char *k, int klen, int plen
     // Set the new header
     WThead *header = GetHeaderWT(nptr, pos);
     header->key_offset = nptr->space_top;
-    header->key_len = (uint8_t)klen;
-    header->pfx_len = (uint8_t)plen;
+    header->key_len = (uint16_t)klen;
+    header->pfx_len = (uint16_t)plen;
 #ifdef WTCACHE
     header->initialized = false;
 #endif
@@ -203,9 +203,9 @@ inline void CopyToNewPageWT(NodeWT *nptr, int low, int high, char *newbase, int 
         WThead *newhead = (WThead *)(newbase + MAX_SIZE_IN_BYTES
                                      - (newidx + 1) * sizeof(WThead));
         strncpy(newbase + top, PageOffset(nptr, oldhead->key_offset), oldhead->key_len);
-        newhead->key_len = (uint8_t)oldhead->key_len;
+        newhead->key_len = (uint16_t)oldhead->key_len;
         newhead->key_offset = top;
-        newhead->pfx_len = (uint8_t)oldhead->pfx_len;
+        newhead->pfx_len = (uint16_t)oldhead->pfx_len;
         top += oldhead->key_len + 1;
     }
 }
@@ -221,8 +221,8 @@ inline void InsertKeyMyISAM(NodeMyISAM *nptr, int pos, const char *k, int klen, 
     // Set the new header
     MyISAMhead *header = GetHeaderMyISAM(nptr, pos);
     header->key_offset = nptr->space_top;
-    header->key_len = (uint8_t)klen;
-    header->pfx_len = (uint8_t)plen;
+    header->key_len = (uint16_t)klen;
+    header->pfx_len = (uint16_t)plen;
     nptr->space_top += klen + 1;
     nptr->size += 1;
 }
@@ -234,9 +234,9 @@ inline void CopyToNewPageMyISAM(NodeMyISAM *nptr, int low, int high, char *newba
         MyISAMhead *newhead = (MyISAMhead *)(newbase + MAX_SIZE_IN_BYTES
                                              - (newidx + 1) * sizeof(MyISAMhead));
         strncpy(newbase + top, PageOffset(nptr, oldhead->key_offset), oldhead->key_len);
-        newhead->key_len = (uint8_t)oldhead->key_len;
+        newhead->key_len = (uint16_t)oldhead->key_len;
         newhead->key_offset = top;
-        newhead->pfx_len = (uint8_t)oldhead->pfx_len;
+        newhead->pfx_len = (uint16_t)oldhead->pfx_len;
         top += oldhead->key_len + 1;
     }
 }
@@ -259,7 +259,7 @@ inline void CopyToNewPageMyISAM(NodeMyISAM *nptr, int low, int high, char *newba
     }
 
 // the k and klen here should always be the fullkey
-inline void InsertKeyPkB(DskManager *dsk, NodePkB *nptr, int pos, const char *k, uint8_t klen, uint8_t plen) {
+inline void InsertKeyPkB(DskManager *dsk, NodePkB *nptr, int pos, const char *k, uint16_t klen, uint16_t plen) {
     // store the full key to disk
     FSeekLongPos(dsk->key_top, dsk->fp);
     fwrite(k, klen + 1, 1, dsk->fp);
